Revert "try using git to fetch registries"
This reverts commit 83ae12900997b46bdec9326bc0391bccd59d2251.

Using git to update registries didn't work very well. It also has
the fundamental problem that the registry origin may be ahead of
what the storage services are ready to serve. So fundamentally we
need to be asking the storage servers what they're ready to serve
we just need a way to validate that it's something that was valid
as the current state of the registry at some point (recently).