Fix registries getting stuck on an old commit
If there are only two upstream StorageServers and they disagree on what
the latest registry URL is, the PkgServer randomly picks one as the
newest commit (contingent on sort order), allowing for the possibility
that it may download new_hash, then old_hash, and stay on old_hash
despite future requests to the storage server consistently returning
new_hash.  This is because if the new registry hash is found on-disk, we
short-circuit all future logic.

This PR instead shortcircuits only the downloading logic, and the hash
replacement logic will still be executed if the file is missing.
We still have the possibility to get confused and serve the older hash,
but as soon as all storage servers agree on a new hash, we are
guaranteed to serve that hash consistently.