Stream in-progress downloads back to the client
Rather than sit and wait for cache misses to make their way from the
StorageServer into the local cache, then start transmitting to the
client, we instead start streaming immediately to the client, before
verification is finished.  This is because we (a) generally trust the
StorageServer to not serve bad data, and (b) will immediately break off
the transmission if the validation fails.  This, coupled with the fact
that the Pkg client should also be verifying allows us to drastically
reduce cold cache latency for large artifacts.