Stream files to the user faster
This commit changes the flow of a cache miss to hit the disk first, then
read back from the disk for both the treehash verification branch and
the stream-unverified-files-to-the-client branch.  This removes one
`tee_task` that would otherwise have linked the speed of tree hash
verification to the speed of serving, as that `tee_task` would start to
block as one of its outputs (the treehash verification consumer) would
begin to get backed up, and fill the `BufferStream`'s entire buffer.  So
instead, we write out to disk first, allowing users to stream off of
disk as fast as `HTTP.get()` can write into that file, essentially using
the disk as an unlimited-size `BufferStream`.  This is mostly just a net
win, although there is a small downside in that if we are getting a bad
resource, we still won't know until we get to the end of verification,
and we may have successfully served the file to many clients in the
meantime.  This is an acceptable downside, as the clients should be
doing their own verification as well.