Teach nginx to reach into the LRU, Revamp deployment, add `k6` testing
Huge update to massively increase PkgServer throughput and reliability
by teaching `nginx` to serve straight out of our file cache.  This gives
an immediate boost to reliability and performance, as `nginx` can serve
static files directly to the client pretty well.

Change deployment to be done directly out of this repository instead of
having most PkgServers use a separate repository.

Add `k6` performance testing framework so that we can more easily dig in
to what is going on inside this machine.