Add registry update watchdog
It appears there are still cases where the registry update thread can
freeze.  We observed several servers within the last week getting
"stuck", where the update task was still alive, there were no errors in
the logs, however they were stuck for over a week.

This may be a deficiency in the downloading code, or even a logic error
somewhere else in the program.  We are adding here a registry update
watchdog that kills the entire server in the event that the registry
update loop does not run for 20 minutes.  This could probably be reduced
significantly, but in the event that things go horribly wrong, we don't
want to be rebooting the server constantly.