Teach PkgServer to serve multiple registry flavors
The storage server now exports two different "flavors" of registry
declarations, `eager` and `conservative`:

- `eager` registries are published as soon as possible, and should track
upstream git repositories quite closely, (in general, it should be less
than two minutes behind the state of the upstream git repository).  The
downside is that the storage server may not have processed some of the
resources in the registry, so if you are a user behind a restrictive
firewall and cannot download resources off of arbitrary servers, we
suggest you use the `conservative` registry.

- `conservative` registries are published only once the StorageServer
has at least tried to process all resources (packages, artifacts,
etc...) within a registry.  It does not guarantee that all resources are
downloadable (after all, there are some package versions that contain
links to artifacts that simply do not exist) but it gives a "best
effort" treehash to use.  The downside is that in the event that a large
number of new resources have been added to the registry recently, it may
take a little while for the storage server to process them all, so the
`conservative` registry flavor may lag behind real-time by a few hours.

Pkg clients that ask for `/registries` are, by default, served a
`conservative` registry, unless one of the following two conditions
hold:

- The Pkg client declares itself to be running CI (via one of the CI
  variables transmitted in the `Julia-Ci-Variables` header, see [0] for
  more), in which case we assume you likely want `eager`.

- The Pkg client transmits a `Julia-Registry-Preference` header set to
  one of the valid flavors, which currently includes `eager` and
  `conservative`.

Now that [1] has been merged into `Pkg.jl`, users can easily set their
own preference header by simply defining the environment mapping
`JULIA_PKG_SERVER_REGISTRY_PREFERENCE=eager`.

[0] https://github.com/JuliaLang/Pkg.jl/blob/bb180de2802a8320241e10721c1ac4eefd3eb27d/src/PlatformEngines.jl#L200-L213

[1] https://github.com/JuliaLang/Pkg.jl/pull/2770