Add `/admin` endpoint
This patch adds an initial `/admin` endpoint that can be accessed with
HTTP Basic authentication. The endpoint is enabled by setting the
environment variable `JULIA_PKG_SERVER_ADMIN_TOKEN_SHA256` to the SHA256
hash of a token that is used as password for the `admin` user.

This patch adds two initial API endpoints:
 - 'GET /admin': Fetch help message about available admin tasks.
 - 'GET /admin/logs': Stream log messages until closing the connection.
   Debug level messages can be enabled by the query parameter
   `level=debug`.