Add initial /admin endpoint.
This patch adds an initial /admin endpoint that can be accessed with
HTTP Basic authentication. The endpoint is enabled by setting
JULIA_PKG_SERVER_ADMIN_TOKEN to a non-empty string. This value can be
used as the password for the 'admin' user in HTTP requests to the /admin
endpoint.

This patch adds four API endpoints:
 - GET /admin: Fetch help message
 - GET /admin/logging: Get the debug-logging state
 - POST /admin/logging: Enable/disable debug logging