Set `JULIA_CPU_TARGET` to make image portable
Some of the servers where we deploy the image doesn't support some of
the instructions that the image is being built with. This means that the
application need to be re-precompiled instead of loading the files
created during image build. This patch sets `JULIA_CPU_TARGET` to the
same value as the generic julia binaries use.

In addition, this patch removes the permission relaxation of
`/depot/compiled` which shouldn't be needed anymore. This was added in
https://github.com/JuliaPackaging/PkgServer.jl/pull/14 to workaround a
similar issue with stale precompilation files.