Expose http ports also when SSL is configured (#192)
This patch changes the nginx configuration to always listen for http
requests on port 8000 and forward them to the upstream package server.

In addition, the docker compose configuration is changed to map port
8001 on the host to port 8000 in the container. This make it possible to
bypass nginx and connect directly to the package server.

To make use of these ports one needs to be on tailnet or use SSH port
forwarding. In the standard configuration ports 8000 and 8001 are both
blocked by the firewall so there are no plain http ports exposed to the
public.