Create nginx log files before starting
This patch make sure that the access_ and error_ logs for nginx are
created before starting nginx. This way we can make sure they are owned
by the regular user. Otherwise, if the files don't exist when nginx
start, they will be created by the nginx master process which runs as
root.

This patch also adjusts the logrotate configuration to create the new
log files with the regular user as owner. This is not necessary as the
file permissions for the newly created logs should match the old files,
but doesn't hurt to explicitly say this in the logrotate configuration.