Run nginx worker process with host uid
This patch adds another layer on top of the docker-nginx-certbot docker
image that creates a new user, nginx-host-uid, with uid and gid matching
that of the regular user running the compose project.

The master nginx process still runs as root, but worker processes run as
the new user. The reason for this is to make sure any files created by
the workers have nicer file permissions.