Cleanup logrotate configuration
This patch removes the hourly cronjob to reopen log files. This is not
needed since logrotate does it in the postrotate section already.
Previously the postrotate section would call the same script as the
hourly cronjob, which resulted in unnecessary reopening of nginx logs.

For the loadbalancer it was previously bugged since postrotate tried to
call a non-existing script (installed as /etc/cron.hourly/loadbalancer
but called in postrotate as /etc/cron.hourly/pkgserver). Since the
script still ran hourly through the cronjob the logs weren't reopened
until 17 past the hour (when hourly jobs run by default). Effectively
this means that nginx used the old log files for a little bit longer.